fix: correct checksum flag assignment and add checksum validation
