+++ gh/gh.go
@@ -6,8 +6,14 @@ import (
 	"compress/bzip2"
 	"compress/gzip"
 	"context"
+	"crypto/md5"  //nolint:gosec
+	"crypto/sha1" //nolint:gosec
+	"crypto/sha256"
+	"crypto/sha512"
+	"encoding/hex"
 	"errors"
 	"fmt"
+	"hash/crc32"
 	"io"
 	"io/fs"
 	"net/http"
@@ -57,6 +63,7 @@ type AssetOption struct {
 	Arch                 string
 	Strict               bool
 	SkipContentTypeCheck bool
+	Checksum             string
 }
 
 func GetReleaseAsset(ctx context.Context, owner, repo string, opt *AssetOption) (*releaseAsset, fs.FS, error) {
@@ -76,6 +83,13 @@ func GetReleaseAsset(ctx context.Context, owner, repo string, opt *AssetOption) 
 	if err != nil {
 		return nil, nil, err
 	}
+
+	if opt != nil {
+		if err := checksum(b, opt.Checksum); err != nil {
+			return nil, nil, err
+		}
+	}
+
 	fsys, err := makeFS(ctx, b, repo, a.Name, []string{a.ContentType, http.DetectContentType(b)})
 	if err != nil {
 		return nil, nil, err
@@ -273,4 +287,63 @@ func matchContentTypes(m, ct []string) bool {
 		}
 	}
 	return false
+}
+
+func checksum(b []byte, c string) error {
+	if c == "" {
+		return nil // No checksum verification needed
+	}
+
+	var (
+		alg  string
+		want string
+	)
+
+	// Check if the format is "algorithm:hash"
+	parts := strings.SplitN(c, ":", 2)
+	if len(parts) == 2 {
+		alg = strings.ToLower(parts[0])
+		want = strings.ToLower(parts[1])
+	} else {
+		// If no alg is specified, try to determine it based on the length of the checksum
+		want = strings.ToLower(c)
+		// Try to match based on length and value
+		switch len(want) {
+		case 8: // CRC32
+			alg = "crc32"
+		case 32: // MD5
+			alg = "md5"
+		case 40: // SHA-1
+			alg = "sha1"
+		case 64: // SHA-256
+			alg = "sha256"
+		case 128: // SHA-512
+			alg = "sha512"
+		}
+	}
+
+	var got string
+	switch alg {
+	case "crc32":
+		got = fmt.Sprintf("%08x", crc32.ChecksumIEEE(b))
+	case "md5":
+		sum := md5.Sum(b) //nolint:gosec
+		got = hex.EncodeToString(sum[:])
+	case "sha1":
+		sum := sha1.Sum(b) //nolint:gosec
+		got = hex.EncodeToString(sum[:])
+	case "sha256":
+		sum := sha256.Sum256(b)
+		got = hex.EncodeToString(sum[:])
+	case "sha512":
+		sum := sha512.Sum512(b)
+		got = hex.EncodeToString(sum[:])
+	default:
+		return fmt.Errorf("unsupported alg: %s", alg)
+	}
+
+	if got != want {
+		return fmt.Errorf("checksum mismatch: expected=%s, calculated=%s", want, got)
+	}
+	return nil
 }
